backend/requirements.txt
aiobotocore==2.15.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aioitertools==0.12.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
black==24.8.0
blinker==1.8.2
blurhash==1.1.4
botocore==1.35.23
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Faker==28.4.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.13.1
